% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnetcoord.R
\name{elnetcoord}
\alias{elnetcoord}
\title{Elastic net regression}
\usage{
elnetcoord(x, y, tol, alpha, lambdagrid)
}
\arguments{
\item{x}{Covariate matrix of \code{n} by \code{p}.}

\item{y}{The depend variable, would be demeaned in the algorithm}

\item{tol}{Tolerance error between two adjacent coordinate descent iterations.
Computed by \code{norm(new.beta - old.beta)}.}

\item{alpha}{Parameter in \code{[0,1]} specifying the relative weight between L1 penalty
and L2 penalty.}

\item{lambdagrid}{Specify the grid to find the most desirable lambda}
}
\description{
Solve the elastic net regression problem.
}
\seealso{
\code{\link[glmnet]{glmnet}}
Can also refer to glmnet for a comparison
}
